+++ pi_controller/cattown/hardware/__init__.py
@@ -1,0 +1,1 @@
+from .hardware import MotionController, MouseState
