+++ pi_controller/requirements.txt
@@ -1,2 +1,4 @@
 numpy==1.26.4
 opencv-python==4.9.0.80
+pygcode==0.2.1
+click==8.1.7
